+++ pkg/sdk/users.go
@@ -5,8 +5,6 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"log"
-	"strconv"
 	"time"
 )
 
@@ -144,18 +142,6 @@ func (row userDBRow) convert() *User {
 		user.LockedUntilTime = row.LockedUntilTime.Time
 	}
 	return user
-}
-
-func handleNullableBoolString(nullableBoolString sql.NullString, field *bool) {
-	if nullableBoolString.Valid && nullableBoolString.String != "" && nullableBoolString.String != "null" {
-		parsed, err := strconv.ParseBool(nullableBoolString.String)
-		if err != nil {
-			// TODO [SNOW-1561641]: address during handling the issue
-			log.Printf("[DEBUG] Could not parse text boolean value %v", nullableBoolString.String)
-		} else {
-			*field = parsed
-		}
-	}
 }
 
 func (v *User) ID() AccountObjectIdentifier {
